requirements.txt
telebot==3.8.1
requests==2.26.0
openpyxl==3.0.9
Pillow==8.3.2
SQLAlchemy==1.4.33